+++ sample-compat/src/app/auth/auth.component.ts
@@ -1,0 +1,67 @@
+import { Component, OnInit, OnDestroy, PLATFORM_ID } from '@angular/core';
+import { AngularFireAuth } from '@angular/fire/compat/auth';
+import firebase from 'firebase/compat/app';
+import { Subscription } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { trace } from '@angular/fire/compat/performance';
+import { Inject } from '@angular/core';
+import { isPlatformServer } from '@angular/common';
+
+@Component({
+  selector: 'app-auth',
+  template: `
+    <p>
+      Auth!
+      {{ (auth.user | async)?.uid | json }}
+      {{ (auth.credential | async)?.additionalUserInfo.isNewUser | json }}
+      <button (click)="login()" *ngIf="showLoginButton">Log in with Google</button>
+      <button (click)="loginAnonymously()" *ngIf="showLoginButton">Log in anonymously</button>
+      <button (click)="logout()" *ngIf="showLogoutButton">Log out</button>
+    </p>
+  `,
+  styles: []
+})
+export class AuthComponent implements OnInit, OnDestroy {
+
+  private readonly userDisposable: Subscription|undefined;
+
+  showLoginButton = false;
+  showLogoutButton = false;
+
+  constructor(public readonly auth: AngularFireAuth, @Inject(PLATFORM_ID) platformId: object) {
+
+    if (!isPlatformServer(platformId)) {
+      this.userDisposable = this.auth.authState.pipe(
+        trace('auth'),
+        map(u => !!u)
+      ).subscribe(isLoggedIn => {
+        this.showLoginButton = !isLoggedIn;
+        this.showLogoutButton = isLoggedIn;
+      });
+    }
+  }
+
+  ngOnInit(): void { }
+
+  ngOnDestroy(): void {
+    if (this.userDisposable) {
+      this.userDisposable.unsubscribe();
+    }
+  }
+
+  async login() {
+    const user = await this.auth.signInWithPopup(new firebase.auth.GoogleAuthProvider());
+    // TODO sign into offline app
+  }
+
+  async loginAnonymously() {
+    const user = await this.auth.signInAnonymously();
+    // TODO sign into offline app
+  }
+
+  logout() {
+    this.auth.signOut();
+    // TODO sign out of offline app
+  }
+
+}