+++ src/compat/performance/public_api.ts
@@ -1,0 +1,3 @@
+export * from './performance';
+export * from './performance.module';
+export * from './performance.service';